[Feat] Big commit. I've moved the services to python, vibe coded it using claude code because I was tired
Instead of bash and yaml for most things, which broke a lot and was inconsistent
I have moved to python which had a dedicated library called proxmoxer to handle
proxmox management. Now the python cli lets us deploy services.
In its current state there's not github actions but it has the ability to discover
and deploy services.